Rework plot.py
Add the ability to create the storage directory if it doesn't exist yet.
Add the ability to run plot.py without specifying a storage directory.
The storage directory is then assumed to be the working directory from
which the script is run.
Reformat the script and add a main function.