Use new event packages in chain model, reformat
Reorder imports using isort package.
Update imports to comply with new package structure.
Use updated event packages that have been slimmed down to not carry the
latest event and the latest entity as these can be determined by caching
the last modeled event in the chain if the chain is sorted. To
accomodate this the event parser returns the chain of events in order of date ascending.
Update the type hints to comply with the new model package names.

See also: #25 #26