Implement objects containing processing conversion
Implement meta objects according to 3 layers of processing:

1. meta layer
First processing step for api objects.
Precompute qualified id's, labels and for meta activities also the start
and end point of the respective activities.

2. PROV atom layer
Convert meta layer objects into simpler representations that only
provide access to the according properties for entities, activities and
agents. Which are id and label for agents and entities and id, label,
start and end for activities.

3. package layer
Package multiple PROV atoms, list of PROV atoms or list of packages into
bigger packages which in turn are used as input for the model
implementations. This conceptualizes the model descriptions used in
gitlab2prov.

See also: #25 #26
Fix: #8