Refactor config.py and fix patch() type conversion
The configparser interprets each option value as a string resulting in
booleans beeing represented by 'False' and 'True' respectivly which
python both treats as truethy as they are non empty-strings. This has
been accounted for by using the strtobool method which converts boolean
string representations to their according boolean value. When
providing a value by a command line flag, the argument parser returns a
real boolean and not a string representation of a boolean resulting in
the type conversion of strtobool() to fail as it only accepts strings
as arguments. This is avoided by first casting to string.

In the future this can be avoided by using the appropriate configparser
read methods for the respective option types.