Skip/Ignore missing parent commits (see #37)
It seems that some commits contain parent sha's of commits that the
gitlab api does not provide when querying for all available commits of a
gitlab project. This commit provides a hotfix to avoid exceptions when
such a parent sha is encountered. The main reason for the issue will
have to be investigated further.