Add status checks for GET reqs, fix url encoding
This commit adds checks for 200 OK status responses for each performed
GET request. If a GET requests answers with anything other than 200 OK
an exception will be raised and the concerned GET request URL will be
diesplayed on the console.

This commit also fixes an issue that made yarl run multiple percentage
encodings on a simple URL encoded path. This has been reversed by
setting the encoded flag for the URL.with_path and other methods to
True.