Add group_by function to gitlab2prov/utils (#42)
The group_by function is used to match page results back to their appropriate request url's.

It is also used to regroup request results such as issue note awards.
For example: The GitlabClient is supposed to return a list of issue note awards for each issue of a project, but it can only request a list of awards for each issue note. To regroup the awards such that there is a list of awards for each issue and not for each issue note, one can use the group_by function and pass a group_size_index that reflects how many issue notes there are on each inbdividual issue.