Hotfix: Remove unclassified event exceptions (#38)
Remove exceptions if no matches or more than one match have been found for a
given system note body. This is only intended as a hotfix and will be
replaced by suffisticated exception handling in the future.

This hotfix returns the properties
    "event": "DEFAULT_EVENT"
    "imported": "DEFAULT_EVENT"
for a system note body that cannot be classified.

This hotfix returns the first match for a given system note body
if more than one match is found.