Add multi-format subcommand to configuration
For configuration files, we try to auto-detect whether or not the
multi-format mode has been chosen.
As soon as multiple serialization formats are detected, we require an
output file to be specified.

For the cli, we add the 'multiformat' subcommand. The command
allows '-f' to accept more than one argument and requires an output file
to be specified.