Avoid code execution at import time
To avoid code execution at import time, the main functionality has been moved to the function "run".
This allows to read the configuration inside of a function and not at import time, which lead to unpredictable errors in the past.
To achieve this, the newly introduced function read_config is used.